requirements.txt
Django==4.1.13
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.2
pillow==10.4.0
django-bootstrap-v5
gunicorn
whitenoise
openpyxl
numpy